updated lab-6

+++ src/main/java/com/app/movietradingplatform/entity/movie/service/MovieService.java
@@ -6,31 +6,69 @@ import com.app.movietradingplatform.entity.movie.Movie;
 import com.app.movietradingplatform.entity.movie.dto.MovieRequest;
 import com.app.movietradingplatform.entity.movie.repository.MovieRepository;
 import com.app.movietradingplatform.entity.user.User;
+import com.app.movietradingplatform.entity.user.UserRoles;
 import com.app.movietradingplatform.entity.user.service.UserService;
-import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.annotation.security.PermitAll;
+import jakarta.annotation.security.RolesAllowed;
+import jakarta.ejb.EJB;
+import jakarta.ejb.LocalBean;
+import jakarta.ejb.Stateless;
 import jakarta.inject.Inject;
-import jakarta.transaction.Transactional;
 import lombok.NoArgsConstructor;
 
 import java.util.*;
 
-@ApplicationScoped
+@LocalBean
+@Stateless
 @NoArgsConstructor()
 public class MovieService {
+    @Inject
     private MovieRepository movieRepository;
     private DirectorService directorService;
     private UserService userService;
 
-    @Inject
-    public MovieService(MovieRepository movieRepository, DirectorService directorService,
-                        UserService userService) {
-        this.movieRepository = movieRepository;
+    @EJB
+    public void setDirectorService(DirectorService directorService) {
         this.directorService = directorService;
+    }
+    @EJB
+    public void setUserService(UserService userService) {
         this.userService = userService;
     }
 
+    @PermitAll
+    public List<Movie> findAll() {
+        return movieRepository.findAll();
+    }
+
+    @PermitAll
+    public Optional<Movie> find(UUID id) {
+        return movieRepository.find(id);
+    }
+
+    @RolesAllowed(UserRoles.ADMIN)
+    public Movie create(Movie movie) {
+        movieRepository.create(movie);
+        return movie;
+    }
+
+    @RolesAllowed(UserRoles.ADMIN)
+    public Movie update(Movie movie) {
+        movieRepository.update(movie);
+        return movie;
+    }
+
+    @RolesAllowed(UserRoles.ADMIN)
+    public void delete(UUID id) {
+        movieRepository.find(id).ifPresent(movieRepository::delete);
+    }
+
+    @RolesAllowed(UserRoles.ADMIN)
+    public void deleteAll() {
+        movieRepository.deleteAll();
+    }
+
     // just for data initialization
-    @Transactional
     public void createWithLinks(MovieRequest request) {
         Movie movie = Movie.builder()
                 .id(UUID.randomUUID())
@@ -51,7 +89,7 @@ public class MovieService {
 
         // link movie in Director object
         if (movie.getDirector() != null) {
-            var director = movie.getDirector();
+            Director director = movie.getDirector();
             if (director.getMovies() == null) director.setMovies(new ArrayList<>());
             director.getMovies().add(movie);
             directorService.update(director);
@@ -59,7 +97,7 @@ public class MovieService {
 
         // link movie in User object
         if (movie.getUser() != null) {
-            var user = movie.getUser();
+            User user = movie.getUser();
             if (user.getOwnedMovies() == null) user.setOwnedMovies(new ArrayList<>());
             user.getOwnedMovies().add(movie);
             userService.update(user);
@@ -67,71 +105,14 @@ public class MovieService {
 
     }
 
-//    public boolean updatePartialWithLinks(PatchSongRequest request, UUID uuid) {
-//        return songRepository.find(uuid).map(song -> {
-//            if (request.getTitle() != null) song.setTitle(request.getTitle());
-//            if (request.getGenre() != null) song.setGenre(request.getGenre());
-//            if (request.getReleaseYear() != null) song.setReleaseYear(request.getReleaseYear());
-//            if (request.getDuration() != null) song.setDuration(request.getDuration());
-//            // re-link artist
-//            if (request.getArtistId() != null) {
-//                var oldArtist = song.getArtist();
-//                if (oldArtist != null) {
-//                    oldArtist.getSongs().removeIf(s -> Objects.equals(s.getId(), song.getId()));
-//                    artistService.update(oldArtist);
-//                }
-//                artistService.find(request.getArtistId()).ifPresent(newArtist -> {
-//                    if (newArtist.getSongs() == null) newArtist.setSongs(new ArrayList<>());
-//                    newArtist.getSongs().add(song);
-//                    artistService.update(newArtist);
-//                    song.setArtist(newArtist);
-//                });
-//            }
-//
-//            // re-link user
-//            if (request.getUserId() != null) {
-//                var oldUser = song.getUser();
-//                if (oldUser != null) {
-//                    if (oldUser.getSongs() != null) oldUser.getSongs().removeIf(s -> Objects.equals(s.getId(), song.getId()));
-//                    userService.update(oldUser);
-//                }
-//                userService.find(request.getUserId()).ifPresent(newUser -> {
-//                    if (newUser.getSongs() == null) newUser.setSongs(new ArrayList<>());
-//                    newUser.getSongs().add(song);
-//                    userService.update(newUser);
-//                    song.setUser(newUser);
-//                });
-//            }
-//
-//            songRepository.update(song);
-//            return true;
-//        }).orElse(false);
-//    }
-//
-    public void deleteWithLinks(UUID movieId) {
-        movieRepository.find(movieId).ifPresent(movie -> {
-            Director director = movie.getDirector();
-            User user = movie.getUser();
-            if (director != null) {
-                if (director.getMovies() != null) director.getMovies().removeIf(
-                        s -> Objects.equals(s.getId(), movie.getId()));
-                directorService.update(director);
-            }
-            if (user != null) {
-                if (user.getOwnedMovies() != null) user.getOwnedMovies().removeIf(
-                        s -> Objects.equals(s.getId(), movie.getId()));
-                userService.update(user);
-            }
-            movieRepository.delete(movie);
-        });
-    }
-
+    @RolesAllowed(UserRoles.ADMIN)
     public List<Movie> findMoviesByDirector(UUID directorId) {
         return directorService.find(directorId)
                 .map(Director::getMovies)
                 .orElse(Collections.emptyList());
     }
 
+    @RolesAllowed(UserRoles.ADMIN)
     public Optional<Movie> findMovieByDirector(UUID directorId, UUID movieId) {
         return directorService.find(directorId)
                 .flatMap(director -> director.getMovies().stream()
@@ -139,7 +120,7 @@ public class MovieService {
                         .findFirst());
     }
 
-    @Transactional
+    @RolesAllowed(UserRoles.ADMIN)
     public Movie createMovieForDirector(UUID directorId, Movie movie) {
         return directorService.find(directorId).map(director -> {
             if (movie.getId() == null) movie.setId(UUID.randomUUID());
@@ -157,7 +138,7 @@ public class MovieService {
         }).orElseThrow(() -> new NoSuchElementException("Director not found"));
     }
 
-    @Transactional
+    @RolesAllowed(UserRoles.ADMIN)
     public Movie updateMovieForDirector(UUID directorId, UUID movieId, Movie updatedMovie) {
         return directorService.find(directorId).flatMap(director -> {
             return movieRepository.find(movieId).map(existingMovie -> {
@@ -176,7 +157,7 @@ public class MovieService {
         }).orElseThrow(() -> new NoSuchElementException("Director or Movie not found"));
     }
 
-    @Transactional
+    @RolesAllowed(UserRoles.ADMIN)
     public void deleteMovieForDirector(UUID directorId, UUID movieId) {
         directorService.find(directorId).ifPresent(director -> {
             movieRepository.find(movieId).ifPresent(movie -> {
@@ -189,6 +170,77 @@ public class MovieService {
                     director.getMovies().removeIf(m -> Objects.equals(m.getId(), movieId));
                 }
                 directorService.update(director);
+
+                // Delete the movie
+                movieRepository.delete(movie);
+            });
+        });
+    }
+
+    @RolesAllowed(UserRoles.ADMIN)
+    public List<Movie> findMoviesByUser(UUID userId) {
+        return userService.find(userId)
+                .map(User::getOwnedMovies)
+                .orElse(Collections.emptyList());
+    }
+
+    public Optional<Movie> findMovieByUser(UUID userId, UUID movieId) {
+        return userService.find(userId)
+                .flatMap(director -> director.getOwnedMovies().stream()
+                        .filter(movie -> Objects.equals(movie.getId(), movieId))
+                        .findFirst());
+    }
+
+    @RolesAllowed(UserRoles.ADMIN)
+    public Movie createMovieForUser(UUID userId, Movie movie) {
+        return userService.find(userId).map(user -> {
+            if (movie.getId() == null) movie.setId(UUID.randomUUID());
+            movie.setUser(user);
+            movieRepository.create(movie);
+
+            // Link the movie in the director's movie list
+            if (user.getOwnedMovies() == null) {
+                user.setOwnedMovies(new ArrayList<>());
+            }
+            user.getOwnedMovies().add(movie);
+            userService.update(user);
+
+            return movie;
+        }).orElseThrow(() -> new NoSuchElementException("User not found"));
+    }
+
+    @RolesAllowed(UserRoles.ADMIN)
+    public Movie updateMovieForUser(UUID userId, UUID movieId, Movie updatedMovie) {
+        return userService.find(userId).flatMap(user -> {
+            return movieRepository.find(movieId).map(existingMovie -> {
+                if (!Objects.equals(existingMovie.getUser().getId(), userId)) {
+                    throw new IllegalArgumentException("Movie does not belong to the specified user");
+                }
+
+                // Update movie details
+                existingMovie.setTitle(updatedMovie.getTitle());
+                existingMovie.setReleaseDate(updatedMovie.getReleaseDate());
+                existingMovie.setGenres(updatedMovie.getGenres());
+                movieRepository.update(existingMovie);
+
+                return existingMovie;
+            });
+        }).orElseThrow(() -> new NoSuchElementException("Director or Movie not found"));
+    }
+
+    @RolesAllowed(UserRoles.ADMIN)
+    public void deleteMovieForUser(UUID userId, UUID movieId) {
+        userService.find(userId).ifPresent(user -> {
+            movieRepository.find(movieId).ifPresent(movie -> {
+                if (!Objects.equals(movie.getDirector().getId(), userId)) {
+                    throw new IllegalArgumentException("Movie does not belong to the specified user");
+                }
+
+                // Remove the movie from the director's movie list
+                if (user.getOwnedMovies() != null) {
+                    user.getOwnedMovies().removeIf(m -> Objects.equals(m.getId(), movieId));
+                }
+                userService.update(user);
 
                 // Delete the movie
                 movieRepository.delete(movie);
